move integration & e2e tests
