+++ tests/e2e/connectors/ccxt/ccxt_integration_test.py
