+++ tests/integration/connectors/ccxt/reader_test.py
